document-ai-for-slack/requirements.txt
# Function dependencies, for example:
# package>=version
certifi==2022.12.7
Flask==2.2.3
slack-bolt==1.16.2
slack-sdk==3.20.0
slackclient==2.9.4
functions-framework==3.2.0
requests==2.28.2
google-cloud-storage
wget==3.2
google-auth==2.16.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
urllib3==1.26.14